+++ integrations/src/sdk/sdkconfiguration.py
@@ -24,9 +24,9 @@ class SDKConfiguration:
     server_idx: int = 0
     language: str = 'python'
     openapi_doc_version: str = '1.0.0'
-    sdk_version: str = '2.1.0'
-    gen_version: str = '2.214.10'
-    user_agent: str = 'speakeasy-sdk/python 2.1.0 2.214.10 1.0.0 wingspan_integrations'
+    sdk_version: str = '3.0.0'
+    gen_version: str = '2.228.1'
+    user_agent: str = 'speakeasy-sdk/python 3.0.0 2.228.1 1.0.0 wingspan_integrations'
     retry_config: RetryConfig = None
 
     def get_server_details(self) -> Tuple[str, Dict[str, str]]: